+++ stryker.conf.js
@@ -1,14 +1,10 @@
 'use strict';
 
-module.exports = config => {
-  config.set({
-    mutator: 'javascript',
-    packageManager: 'npm',
-    reporters: ['html', 'clear-text', 'progress'],
-    testRunner: 'command',
-    commandRunner: { command: 'npm run test:unit' },
-    transpilers: [],
-    coverageAnalysis: 'all',
-    mutate: ['lib/*.js', '!lib/speed.js', '!lib/report.js', '!lib/utils.js'],
-  });
+module.exports = {
+  packageManager: 'npm',
+  reporters: ['html', 'clear-text', 'progress'],
+  testRunner: 'command',
+  commandRunner: { command: 'npm run test:unit' },
+  coverageAnalysis: 'all',
+  mutate: ['lib/*.js', '!lib/speed.js', '!lib/report.js', '!lib/utils.js'],
 };