+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased][unreleased]
 
+## [0.8.0][] - 2021-01-14
+
 ### Added
 
 - `metatests.measure()` function to perform microbenchmarks with
@@ -563,7 +565,8 @@ and this project adheres to
 
 - The first version of the `metatests` package.
 
-[unreleased]: https://github.com/metarhia/metatests/compare/v0.7.2...HEAD
+[unreleased]: https://github.com/metarhia/metatests/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/metarhia/metatests/compare/v0.7.2...v0.8.0
 [0.7.2]: https://github.com/metarhia/metatests/compare/v0.7.1...v0.7.2
 [0.7.1]: https://github.com/metarhia/metatests/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/metarhia/metatests/compare/v0.6.6...v0.7.0